いったん動く状態
